+++ src/components/Display.js
@@ -12,7 +12,7 @@ const Display = props => {
             <p id="timer-label">{props.isSesh?'session':'break'}</p>
             <p id="time-left">{props.value}</p>
             <button id="start_stop" onClick={props.handleTimerAction}>{props.isPause?'start':'pause'}</button>
-            <button id="reset" onClick={props.handleReset}></button>
+            <button id="reset" onClick={props.handleReset}>reset</button>
         </div>
     )
 }